storage: split collection_mgmt out into module
